+++ inc/commondbtm.class.php
@@ -1067,7 +1067,7 @@ class CommonDBTM extends CommonGLPI {
          if (($name = $this->getName()) == NOT_AVAILABLE) {
             //TRANS: %1$s is the itemtype, %2$d is the id of the item
             $this->fields['name'] = sprintf(__('%1$s - ID %2$d'),
-                                            $this->getTypeName(1), $this->fields['id']);
+                                            $this->getAssetTypeName(1), $this->fields['id']);
          }
          $display = (isset($this->input['_no_message_link'])?$this->getNameID()
                                                             :$this->getLink());
@@ -1341,7 +1341,7 @@ class CommonDBTM extends CommonGLPI {
          } else {
             //TRANS: %1$s is the itemtype, %2$d is the id of the item
             $this->fields['name'] = sprintf(__('%1$s - ID %2$d'),
-                                            $this->getTypeName(1), $this->fields['id']);
+                                            $this->getAssetTypeName(1), $this->fields['id']);
          }
 
          if (isset($this->input['_no_message_link'])) {
@@ -2437,10 +2437,10 @@ class CommonDBTM extends CommonGLPI {
                ]
             );
          } else if ($this->isNewID($ID)) {
-            $nametype = $params['formtitle'] !== null ? $params['formtitle'] : $this->getTypeName(1);
+            $nametype = $params['formtitle'] !== null ? $params['formtitle'] : $this->getAssetTypeName(1);
             printf(__('%1$s - %2$s'), __('New item'), $nametype);
          } else {
-            $nametype = $params['formtitle'] !== null ? $params['formtitle'] : $this->getTypeName(1);
+            $nametype = $params['formtitle'] !== null ? $params['formtitle'] : $this->getAssetTypeName(1);
             if (!$params['noid'] && ($_SESSION['glpiis_ids_visible'] || empty($nametype))) {
                //TRANS: %1$s is the Itemtype name and $2$d the ID of the item
                $nametype = sprintf(__('%1$s - ID %2$d'), $nametype, $ID);
@@ -4256,7 +4256,7 @@ class CommonDBTM extends CommonGLPI {
 
                case "itemtypename" :
                   if ($obj = getItemForItemtype($value)) {
-                     return $obj->getTypeName(1);
+                     return $obj->getAssetTypeName(1);
                   }
                   break;
 
@@ -4585,13 +4585,13 @@ class CommonDBTM extends CommonGLPI {
       if ($result) {
          echo "<div class='center'><table class='tab_cadre'>";
          if ($add) {
-            echo "<tr><th>" . $item->getTypeName(1)."</th>";
+            echo "<tr><th>" . $item->getAssetTypeName(1)."</th>";
             echo "<th>".__('Choose a template')."</th></tr>";
             echo "<tr><td class='tab_bg_1 center' colspan='$colspan'>";
             echo "<a href=\"$target_blank\">".__('Blank Template')."</a></td>";
             echo "</tr>";
          } else {
-            echo "<tr><th>".$item->getTypeName(1)."</th>";
+            echo "<tr><th>".$item->getAssetTypeName(1)."</th>";
             if (Session::isMultiEntitiesMode()) {
                echo "<th>".__('Entity')."</th>";
             }