+++ inc/commondropdown.class.php
@@ -98,7 +98,7 @@ abstract class CommonDropdown extends CommonDBTM {
     *
     *  @since version 0.85
    **/
-   static function getMenuContent() {
+   public function getMenuContent() {
       global $CFG_GLPI;
 
       $menu = [];