+++ inc/config.class.php
@@ -71,7 +71,7 @@ class Config extends CommonDBTM {
     *
     *   @since version 0.85
    **/
-   static function getMenuContent() {
+   public function getMenuContent() {
       global $CFG_GLPI;
 
       $menu = [];