+++ inc/plugin.class.php
@@ -78,7 +78,7 @@ class Plugin extends CommonDBTM {
     *
     * @since version 0.85
    **/
-   static function getMenuContent() {
+   public function getMenuContent() {
       global $CFG_GLPI;
 
       $menu = [];