+++ inc/rule.class.php
@@ -195,7 +195,7 @@ class Rule extends CommonDBTM {
     *
     *  @since version 0.85
    **/
-   static function getMenuContent() {
+   public function getMenuContent() {
       global $CFG_GLPI;
 
       $menu = [];